+++ src/commands/Minion/create.ts
@@ -3,7 +3,7 @@ import { Bank } from 'oldschooljs';
 
 import { BotCommand } from '../../lib/BotCommand';
 import { Time } from '../../lib/constants';
-import Createables from '../../lib/createables';
+import Createables from '../../lib/data/createables';
 import { UserSettings } from '../../lib/settings/types/UserSettings';
 import { SkillsEnum } from '../../lib/skilling/types';
 import {
